refactor: ActivityCard & Like
